# CRC24 performance improvement (32 bits in parallel)