@randombit 2017-10-29
(Reviewed in #1280)

3.5x faster on my machine, this is great. Thanks again